Rename ExceptionHandlerModel
